+++ R/improve_beu_from_rule_dt.R
@@ -1,32 +1,54 @@
-#' improve beu from rules
+#' update beu from rules
 #'
 #' update beumc values based on rules
 #'
 #' @param vri_bem sf object that represent VRI (vegetation ressource inventory) features
 #' @param rules_dt data.table object that contains rule to apply to vri_bem to update beumc
 #' @return sf object
-#' @import rlang, data.table
+#' @import rlang
+#' @import data.table
 #' @export
-improve_beu_from_rule_dt <- function(vri_bem, rules_dt) {
+update_beu_from_rule_dt <- function(vri_bem, rules_dt) {
 
-  # create expression for BGC_ZONE rule
-  rules_dt[, BGC_ZONE_expr := fcase(!is.na(BGC_ZONE), paste0("BGC_ZONE == '", BGC_ZONE, "'"),
-                                    default = "TRUE")]
+  setDT(vri_bem)
 
-  # create expression for BGC_SUBZON rule
-  # TODO shortcut of using hardcoded start and stop for substr only works for one letter condition
-  rules_dt[, BGC_SUBZON_expr := fcase(grepl("^CONTAINS", BGC_SUBZON), paste0("grepl('",substr(BGC_SUBZON, 10, 10),"', BGC_SUBZON)"),
-                                      grepl("^DOES NOT CONTAIN", BGC_SUBZON), paste0("!grepl('",substr(BGC_SUBZON, 10, 10),"', BGC_SUBZON)"),
-                                      !is.na(BGC_SUBZON), paste0("BGC_SUBZON == '", BGC_SUBZON, "'"),
-                                      default = "TRUE")]
+  # find rule columns except tree rules
+  rules_dt_names <- names(rules_dt)
+  length_rules_dt_names <- length(rules_dt_names)
+  which_name_input <- which(rules_dt_names == "INPUTS")
+  which_name_output <- which(rules_dt_names == "OUTPUTS")
 
-  # create expression for BGC_VRT rule
-  rules_dt[, BGC_VRT_expr := fcase(!is.na(BGC_VRT), paste0("BGC_VRT == ", BGC_VRT),
-                                   default = "TRUE")]
+  tree_list_var <- grep("^TREE_RL_SP_CD_[0-9]$", names(rules_dt))
+  tree_pct_var <- grep("^TREE_RL_SP_PCT_[0-9]$", names(rules_dt))
+  rule_columns <- setdiff((which_name_input + 1):(which_name_output - 1), c(tree_list_var, tree_pct_var))
+
+  # create expession for all rule column except for tree rules
+
+  # for non tree rules there is 4 possibilities
+  # the CONTAINS keyword creates a rule that check if the variable contains the specified value
+  # the DOES NOT CONTAINS keyword creates a rule that check if the variable does not contains the specified value
+  # a list of values separate by commas creates a rule that check if the variable is in the list of values
+  # a single value creates a rule that check if the variable is equal to the value
+
+  for (column in rule_columns) {
+    column_name <- rules_dt_names[column]
+    column_name_expr <-  parse_expr(column_name)
+    rules_dt[, paste0(rules_dt_names[column], "_expr") := fcase(grepl("^CONTAINS", eval(column_name_expr)), var_contains(var_name = column_name, rule = eval(column_name_expr)),
+                                                                grepl("^DOES NOT CONTAIN", eval(column_name_expr)), var_does_not_contains(var_name = column_name, rule = eval(column_name_expr)),
+                                                                grepl(",", eval(column_name_expr)), var_in_list(var_name = column_name, rule = eval(column_name_expr)),
+                                                                !is.na(eval(column_name_expr)), var_equal_value(var_name = column_name, rule = eval(column_name_expr)),
+                                                                default = "TRUE")]
+
+  }
 
   # create expression for each pairs of tree list and percentage
-  tree_list_var <- grep("^TREE_RL_SP_CD_[0-9]$", names(rules_dt), value = T)
-  tree_pct_var <- grep("^TREE_RL_SP_PCT_[0-9]$", names(rules_dt), value = T)
+
+  # for tree rules there is two possibilities
+  # a list of values creates a rule that check if the total sum of percentage of any species in the list across all the pairs of species/pct variables is within the range in the corresponding pct rule column
+  # a list of values separated by a > or a < sign creates a rule that check of the total sum of percentage of the species in the left-hand-side of the sign across all the pairs of species/pct variables is greater or less than the percentage for the species in the right-hand-side
+
+  tree_list_var <- rules_dt_names[tree_list_var]
+  tree_pct_var <- rules_dt_names[tree_pct_var]
 
   vri_bem_species_var <- grep("^SPEC_CD_[0-9]$", names(vri_bem), value = T)
   vri_bem_pct_var <- grep("^SPEC_PCT_[0-9]$", names(vri_bem), value = T)
@@ -40,25 +62,29 @@ improve_beu_from_rule_dt <- function(vri_bem, rules_dt) {
 
   }
 
-  # create expression for SOIL_MOISTURE_REGIME_1 rule
-  rules_dt[ , SOIL_MOISTURE_REGIME_1_expr := fcase(!is.na(SOIL_MOISTURE_REGIME_1), paste0("SOIL_MOISTURE_REGIME_1 %in% ", convert_rule_list_to_string_vector(SOIL_MOISTURE_REGIME_1)),
-                                                   default = "TRUE")]
-
-  # create expression for slope mod
-  rules_dt[ , SLOPE_MOD_expr := fcase(!is.na(SLOPE_MOD), paste0("SLOPE_MOD %in% ", convert_rule_list_to_string_vector(SLOPE_MOD)),
-                                                   default = "TRUE")]
-
   # create total expression
-  rules_dt[ , total_expr := parse_exprs(do.call(paste, c(.SD, list(sep = " & ")))), .SDcols = c("BGC_ZONE_expr", "BGC_SUBZON_expr", "BGC_VRT_expr", paste0(tree_list_var, "_expr"), "SOIL_MOISTURE_REGIME_1_expr", "SLOPE_MOD_expr")]
+  rules_dt[ , total_expr := parse_exprs(do.call(paste, c(.SD, list(sep = " & ")))), .SDcols = c(paste0(rules_dt_names[rule_columns], "_expr"), paste0(tree_list_var, "_expr"))]
 
-  # apply total expr on vri_bem and update BEUMC based on rules result
+  # apply total expr on vri_bem and update output columns based on rules result
   for (rule in 1:nrow(rules_dt)) {
     which_lines <- vri_bem[, which(eval(rules_dt[["total_expr"]][[rule]]))]
-    # TODO find out which BEUMC to replace between the 3 deciles
-    set(vri_bem, i = which_lines, j = "BEUMC_test", value = rules_dt[["BEUMC"]][rule])
+    for (output_col in (which_name_output + 1):length_rules_dt_names) {
+      # TODO find out which BEUMC to replace between the 3 deciles
+      set(vri_bem, i = which_lines, j = rules_dt_names[output_col], value = rules_dt[[rules_dt_names[output_col]]][rule])
+    }
   }
 
-  return(vri_bem)
+  return(st_as_sf(vri_bem))
+}
+
+var_contains <- function(var_name, rule) {
+  n_char_rule <- nchar(rule)
+  paste0("grepl('",substr(rule, 10, n_char_rule),"', ", var_name, ")")
+}
+
+var_does_not_contains <-  function(var_name, rule) {
+  n_char_rule <- nchar(rule)
+  paste0("!grepl('",substr(rule, 18, n_char_rule),"', ", var_name, ")")
 }
 
 convert_rule_list_to_string_vector <- function(rule_list) {
@@ -66,6 +92,14 @@ convert_rule_list_to_string_vector <- function(rule_list) {
   res <- paste0("c('", sapply(strsplit(gsub(" ", "", rule_list), split = ",|>|<"), function(x) paste0(x, collapse = "','")), "')")
   res[which_na] <- ""
   return(res)
+}
+
+var_in_list <- function(var_name, rule_list) {
+  paste0(var_name, " %in% ", convert_rule_list_to_string_vector(rule_list))
+}
+
+var_equal_value <- function(var_name, rule_value) {
+  paste0(var_name," == '", rule_value, "'")
 }
 
 sum_pct_for_species_in_list <-  function(tree_list, vri_bem = NULL, species_var = NULL, pct_var = NULL) {
@@ -78,7 +112,7 @@ sum_pct_for_species_in_list <-  function(tree_list, vri_bem = NULL, species_var 
 
   species_list <- convert_rule_list_to_string_vector(tree_list)
 
-  for (i in seq_along(vri_bem_species_var)) {
+  for (i in seq_along(species_var)) {
     if (i == 1) {
       string_expression <- paste0("(", species_var[i], " %in% ", species_list, ") * as.numeric(", pct_var[i], ") ")
     } else {